Cocoa: Implement our own NSApplication subclass
We have avoided doing so up till now, since we cannot always
know if we will be able to use it. If some 3rd-party application
creates NSApplication before Qt, our subclass will never be
used (because of the singleton pattern that NSApplication follows).
However, in most cases, Qt will be used in standalone applications,
or the 3rd-party application will not subclass NSApplication. And
in those cases, we can make certain functionallity more robust.

Rev-By:msorvig