Fixed drawing QPixmaps on QGLWidgets on different X11 screens.
Don't try to use the texture_from_pixmap extension when drawing a pixmap
from one X11 screen onto a different X11 screen in the GL 2 engine.
For that to work, the pixmap will have to be re-created on the new
screen, which is exactly what the default bind method does (via QImage
conversion).

Task-number: QTBUG-8054
Reviewed-by: Kim