Factored epocRoot implementation out of qmake
This function is now implemented in its own source file, rather
than being embedded within the Symbian qmake generator.

The motivation for this is to allow code to be shared between
qmake and configure - the latter needs to determine the epoc
root path in order to perform feature detection on Symbian SDKs.

Reviewed-by: Miikka Heikkinen