Blinking cursors are 2 pixels wide on Mac OS X/Cocoa.
Blinking cursors drawn in positions other than zero, can become 2 pixel
wide. This is caused by a rendering bug in the paint engine used by
Cocoa. If a fillRect() is called in a nox pixel boundary this engine
draws them 2 pixels wide instead of one. So make sure this is always on
a pixel boundary.

Task-number: QTBUG-8100
Reviewed-by: Trond