Wrong cursor shown by the parent window after setOverrideCursor().
When calling QApplication::setOverrideCursor() and then showing a modal
dialog, the parent window of the dialog reverts to the default cursor.
At this point the parent window is in a disabled state, Windows uses the
default cursor for the class. We need to override this behavior by
handling WM_SETCURSOR.

Task-number: QTBUG-6525
Reviewed-by: Denis Dzyubenko