Move http resume code into http backend.
Adds two new virtual functions to QNetworkAccessBackend:

bool canResume() const; and
void setResumeOffset(quint64)

canResume() is used to query the existing backend if resuming is
supported for the current request.

setResumeOffset() is used to set an additional offset which should be
applied when the backend posts the request.

The default implementation of canResume() returns false (resuming not
supported). The default implementation of setResumeOffset() does
nothing.