Replaced the numCopies()/setNumCopies() and actualNumCopies() API.
Obsoleted the numCopies()/setNumCopies()/actualNumCopies() with
copyCount()/setCopyCount() and supportsMultipleCopies(). The
old numCopies() was confusing since it returned the number
of times a job had to be printed to get the copies you wanted.
Now copyCount() and setCopyCount() are symmetric, and always returns
the number of copies the user want to print.

Task-number: QTBUG-5605
Reviewed-by: Kim