Prevent stale QWidget pointers in QCocoaView
QCocoaView has two "back" pointers ot its owning QWidget
and the corresponding QWidgetPrivate. The lifetime of the
QCocoaView might be longer and that of the QWidget if Cocoa
itself holds references to it. In addition accessing a
QWidget that is being destroyed is not safe.

Set both pack pointers to NULL in ~QWidget. This prevents
following stale pointers in the QCocoaView callbacks. This
also means that we'll see the NULL pointers, add gurads for
that.