Mac: submenu shows up at the wrong position
If one opens a menu with a submenu, the submenu will be shown at a
wrong position on screen. The reason is that upon show, the top menu
will be centered on screen. But the child menu will be shown first, and
tries to center of the parent that has not yet been shown. So moving the
top menu upon show will fool the subwindow.
This patch will move the center code into the setGeometry_sys function
so that any resize on the window before it has been shown will re-center
the window correctly.

Task-number: QTBUG-2196
Reviewed-by: prasanth