Fixed missing repaints on S60 3.1.
The bug happened when a paint event came in from Symbian, but the
painted area was bigger than the exposed rect being passed in by
Symbian. In these cases we would lose updates to the area outside.

Fixed by calling Draw() once more if we detect that we painted a
larger area to the backing store.

It is still a mystery why this worked on 5.0 though... :-P

AutoTest: QWidget passed
RevBy:    Jason Barron
Task:     QTBUG-8200