Partial upgrade packages for QtCore and QtGui
When debugging or benchmarking changes in QtCore / QtGui, it is
inconvenient to reinstall the whole of Qt.
This allows 'make sis' to generate a partial upgrade sis files,
which can be installed on top of an existing Qt installation.

This relies on binary compatibility, so you should be using def files.
Self-signed partial upgrade can only install on top of self-signed Qt.

Reviewed-by: Miikka Heikkinen