Fix a deadlock in kqueue implementation of QFileSystemWatcher
When adding and removing lots of files, the kqueue implementation's
pipe between the worker thread and the main interface would fill up,
causing the thread adding/removing files to block even though the
worker thread was no longer running.

The kqueue interface gives us information about how much data is
available on the pipe (the data member of the kevent struct), so we
should use this to make sure we always empty the pipe. We then iterate
over all the command bytes and make sure we interpret them correctly.

Reviewed-by: Prasanth Ullattil