Fix broken image loading on windows.
Make sure the url is valid and we could find the file in the compressed
help, otherwise pass on an error message. We might have got garbage in
the first patch, so the isEmpty() check was not sufficient.

Do not integrate in master, since it was fixed already there.

Task-number: QTBUG-8082, QTBUG-8248
Reviewed-by: ck