Fix Thai text on Windows 7
Segoe UI, which is the default ui font on Windows 7, contains the glyph
for a single character in the thai unicode range. This is thus detected
as supporting the Thai script even though it only contains this one
character, and as a result all thai text would appear as boxes on
Windows 7 unless you changed the font. The patch is a temporary hack
to work around the problem until a proper fix can be made.

Reviewed-by: Lars Knoll