optimization: get rid of QString::fromUtf16() usage, part 2
QString::fromUtf16() is slow - it does a BOM check and optionally byte
swapping, which is utterly irrelevant when converting internal data
structures which are raw utf16 in host byte order. so replace it with
QString::fromUnicode() where possible (which seems to be everywhere).

the reasoning is the same as in commit e0fda52f, so not getting further
reviews.

Reviewed-by: denis
Reviewed-by: joao