Fixes crash when destroying a QGraphicsItem.
When clearing the focus on an item, the focus was previously passed to
the closest parent focus scope (if any). But the focus should go to the
closest parent focus scope only if the item being cleared of the focus
is itself a focus scope. This incorrect behavior leaded to invalid
pointers in QGraphicsItem under specific circumstances, eventually
leading to a crash when destroying the item.

Auto-test included.

Task-number: QT-2649
Reviewed-by: Andreas Aardal Hanssen