Make sure height of glyph cache is a power of two
Making the size of the glyph cache a power of two can potentially
have performance benefits on hardware accelerated graphics systems.
Since it will also decrease the number of times the cache has to be
copied, and therefore implies a speed-up for all paint engines, it has
been implemented for all of them in general.

Reviewed-by: Gunnar