Avoid calling out to public API in the QtScript implementation
There's no reason to construct QScriptValues when JSC::JSValues
can be operated on directly. A benchmark showed that ~10% of the
time for reading a QObject property from a script was spent just
creating and destroying QScriptValue temporaries. It's time to
finally get rid of this potential bottleneck, not just in the
QObject integration but everywhere.

This change refactors the code so that all internal operations
are performed on JSC::JSValue (most importantly, conversion
from/to Qt types), and the public API functions are just thin
wrappers around these. E.g. instead of doing

enginePrivate->scriptValueFromJSCValue(jsValue).toQObject()

the implementation now does

QScriptEnginePrivate::toQObject(jsValue)

Other operations are delegated to the engine implementation
in similar style.

Task-number: QTBUG-8304
Reviewed-by: Jedrzej Nowacki