optimize queue-like structures
a list to which we append after we took something from the beginning
is most likely a queue, and it seems unlikely that we would suddenly
start prepending to it. consequently, optimizing the prepending case
by leaving the first third of the allocation free just increases the
number of times the array needs to be shifted down.

Reviewed-by: joao