Virtual keyboard can't be closed anymore after tapping the text area
On Symbian^3, the virtual keyboard cannot be closed anymore
when the editor area where the text is displayed is tapped and
autocompletion is suggesting a word currently.

This is caused by the fragile native FEP manager which assumes that
client must report certain event (EAknCursorPositionChanged), after
user interaction with the editor area. Otherwise the native FEP
manager is in dead-lock, until somebody reports this event to it.

Task-number: QTBUG-7828
Reviewed-by: axis