Remote lib extensions for Windows Mobile device power operations.
Function to change the power mode and to reset the
device were added. The reset is needed to put the
device in a defined state after an error has occurred.
The unattended power mode ensures that the reset can be
effected and that the device will not go to sleep.

Reviewed by: Joerg