Ensure that posted events are sent on Windows
Commit f21d183 introduced a GetMessage hook that would try to post
a message to the Windows queue if the hook received an event that was
not the WM_QT_SENDPOSTEDEVENTS message.

However, the logic used is flawed, and would never post the message
unless we received a message for a window that was not the event
dispatcher's internal window. Timer messages DO go to the internal
window, and the last handled timer message SHOULD trigger the hook
to post a new WM_QT_SENDPOSTEDEVENTS, but due to the flawed logic, it
would not.

The most notable side effect of this bug is that regular repaints and
animations would not become visible unless the user moved the mouse
or used the keyboard.

Task-number: QTBUG-7728
Reviewed-by: Andreas Aardal Hanssen

Fix verified by 2 users in the Jira report mentioned above.