Cocoa: event dispatcher eats mouse events
It turns out that the event dispatcher did not flush queued user input
events under some circumstances (when adding the exec-flag to
processEvents). And this caused problems in the QML-editor in creator
regarding focus frames. This patch makes sure that we always flush
the queued user input events when calling processEvents.

Task-number: QTBUG-8274
Reviewed-by: Prasanth
Reviewed-by: cduclos