create temporaries more intelligently
we already know that we are dealing with a movable type, so it is safe
to keep a temporary copy of the raw data instead of going through the
copy c'tor.
this way we save a pretty useless ref()/deref() pair for each and every
insertion of an implicitly shared type into a QList.

the QtPodForSize class is somewhat ugly. we should use QIntegerForSize,
but that's not possible without having separate template
specializations, which is not possible without some bigger changes to
the qt type info system. it will be beautified in due time. :)

Reviewed-by: joao