Enable window size caching on Symbian NGA variants.
Enabling this flag saves us the round-trip to WSERV whenever
RWindow::Size() is called because the size is cached on the client
side. This can improve performance because functions like
eglSwapBuffers() call Size() to see if the window size has changed and
without the cache this introduces an extra IPC call for every frame.

Task-number: QT-2849
Reviewed-by: Iain