add possibility to request project recursion from within a pro file
this is the same as specifying -r on the command line, but it can be set
in pro and spec files for cases where no recursion would plain not work.

the implementation is via a new option() instruction which at some point
will be used to set other "modes of operation" as well. for now, only
the "recursive" option is recognized.

Reviewed-by: mariusSO