QSortFilterProxyModel: Sorting occured unnecessarily when the dynamicSortFilter is turned off
We should not sort when inserting items if the dinamicSortFilter flag
is set to false.
Note that some of the test used to rely on the fact that it was sorted.
Those test have been fixed.

The patch has been contributed to us in the task.

Task-number: QTBUG-7716
Reviewed-by: Thierry
(cherry picked from commit 895b9bedc3746723f6c77754df3c428dbc0661d3)