Factored readRegistryKey implementation out of qmake
This function is now implemented with its own header and source
files, rather than being embedded within the MSVC qmake generator.

The motivation for this is to allow code to be shared between
qmake and configure, both of which query the registry when built for
Windows.

Reviewed-by: Miikka Heikkinen
(cherry picked from commit 5254184c07b4da800e29000a251ed2a026bd2be5)

Conflicts:

	qmake/Makefile.unix