Implemented recursive "sis" target and enabled sis_targets feature.
The sis target will depend on the main build target of each project,
which should make it possible to simply run "make sis" in the root of
a clean tree (after configure, of course), and have sis packages
generated for everything. It works at least for Qt.

This required some changes to the dependency generation in Qt, to
make sure that s60installs builds its package after all of Qt, and
that fluidlauncher has all the required files before building a
package.

In addition, all the sis target code was moved into its own qmake
feature file, called sis_targets. It is currently enabled by default
for Symbian.