Regression: QGraphicsScene::render fails to render the entire scene correctly.
This only happened with items that had either negative width or height
(boundingRect().width()|height()). Problem was that in case of not having
an exposed region (drawing items from QGraphicsScene::render), we simply
checked whether the bounding rect was empty or not. This is fine, however
we have to normalize the rect first.
(Note that QRegion::intersects(rect) always normalizes the rect, so
that's why this use case broke only when calling QGraphicsScene::render).

Auto-test included.

Task-number: QTBUG-7775
Reviewed-by: yoann