Remove comment in drawStaticTextItem() in GL engine
Removing this comment, as a decision was made. QStaticText is an
optimization class. Forcing a fallback for all fonts over a certain
size would be to intrusive. People should get to decide for themselves
whether to cache the text or not when using QStaticText. They can always
fall back to drawText() themselves.

Reviewed-by: TrustMe