Improvements to itemview keypad navigation in S60.
The logic used in this commit is partially taken from sliders.

This commit makes it possible to interact (change row or column)
in itemview even itemview does not have editFocus. Interacting without
editFocus is enabled when it is not possible to keypad navigate to
reuqested direction. In addition if keypad navigation to any direction is
not possible (i.e there is only one listwidget on screen), there is no
sense to add "done" softkey to get out of edit focus.

Task-number: QTBUG-4802
Reviewed-by: Alessandro Portale