Fix SymbianMakefileGenerator::absolutizePath for clean builds
QFileInfo::isDir() will not return correct value if path doesn't exist
like it doesn't in case of clean builds. Turned the check around
and used isFile() instead. If a file is given as origPath, it is
assumed to exist, because it was explicitly given with HEADERS variable
and qmake will give warning if nonexisted file is given there.

Reviewed-by: Janne Koskinen