ActiveQt Internet Explorer component causes Desktop Icons to flicker.
This happens only if IE is embedded in an frameless window. The repaint
caused by calling EnableModeless() is making the desktop icons flicker.
This repaint generated via ActiveQt will be ignored by QtGui. This is
done by checking if the window style is already in the required state
or not.

Task-number: QTBUG-8355
Reviewed-by: Denis