Improve lookup speed of QMetaType::type() for normalized types
Two improvements are made:
1) Store the length of the built-in types to avoid calling
strcmp() when there's no way the types will match; and
2) Speculate on the input already being normalized.

When the type name is already in normalized form (which we
believe to be the common case), normalizing is wasteful
(allocating a bytearray that's going to contain exactly the
same string as the original input).

Use the same pattern as used in QObject::connect(), which
speculates on the signature already being normalized; only
if the lookup fails, the signature is normalized and looked
up again.

For the QMetaType::builtinTypes benchmark, this improves
performance by roughly 4x.

For the QMetaType::builtinTypesNotNormalized benchmark, the
performance is half of that before, because the lookup is
done twice. But we choose to optimize for the already-normalized
case. This makes a big difference in the QtScript QObject
bindings, for example.

Reviewed-by: Harald Fernengel