Added fullscreen support to softkeys in Symbian.
This commit enables the following two features:

1. Developer can make softkeys visible in fullscreen widget by setting
the Qt::WindowSoftkeysVisibleHint window flag. This flag implememts
intermediate mode for maximized and fullscreen modes. In maximized mode
both statuspane and softkeys are visible, in fullscreen mode with
WindowSoftkeysVisibleHint flag, only the softkeys are visible and
in normal fullscreen mode both statuspane and softkeys are invisible.
This feature was requested by QTBUG-5171.

2. Developer can make softkeys to respond to the key events even the
softkeys are invisible. This means that when widget with
Qt::WindowSoftkeysRespondHint window flag is shown in fullsreen, the
softkey HW key events are routed to invisible softkeys and softkeys
trigger the action associated to pressed softkey button. If the flag
is not set, the key event will be passed to application/focused widget
normally and softkey actions are not triggered. This feature was
requested in QTBUG-4564.

Both new flags are by default off.

In addition, the softkey example is updated to demonstrate the new
window flags. The commit also contains some code style fixes applied by
my editor automatically to changed files.

Task-number: QTBUG-5171
Task-number: QTBUG-4564

Reviewed-by: Jason Barron