Added orientation change support for show/showNormal widgets in S60.
Earlier the widget normalGeometry was calculated once when widget was
made visible first time. Because geometry was not recalculated
the widget displayed with show/showNormal did not change its location
or size as a result of orientation change. This lead to unefficient
screen usage.

This commit makes widget to recalculate its normalGeometry as a result
of orientation change, if developer has not set the widget position or
size explicitly.

Task-number: QTBUG-8551
Reviewed-by: Jason Barron