qdoc: Second attempt, QML doc stuff.
This change caused a crash on several platforms before,
but not on mine, of course. It outputs a list of new
QML elements on the What's New in 4.7 page, and it
outputs the "Inherited by" list on each QML elemnent
page for elements that are inherited by other elements.
I can't see why it should crash anything, but I have
included some debug output. I think the crash might
have something to do with QList<T>, which qdoc3 uses
heavily. If it crashes for you, please get a stack
trace and the debug output.