Mac: static getFont functions in QFontDialog does not work
This bug is visible if the static funtions are used before
calling qApp->exec().
The reason is that the static functions still use the old
code path, rather than the new one (that will end up creating
a QFontDialog and call exec on it). Just using the new style
will fix the problem.

Task-number: QTBUG-7371
Reviewed-by: cduclos