make the value of QMAKE_QMAKE somewhat less magic
the generators change the value of QMAKE_QMAKE, so it is unwise to
"redirect" it to a hidden builtin which is reset each time.

in particular, this fixes qmake generating makefiles without an absolute
path to qmake itself - the initial quoting of the filename will make the
variable "real", so contains() will start working for it.