Make QLabel::text a reloadable property
In this way we can correctly load / reload a label in case it contains
a rich text with a reference to the image taken from resources.
Since QLabel::setText() ignores the call in case we try to set the same
text, we force the reload by setting the empty string in between.
For performance reasons we do it only in cases when the text value
contains :/ (only in this case it might have a reference
to the resources).

Reviewed-by: Friedemann Kleint <Friedemann.Kleint@nokia.com>
Task-number: QTBUG-8347