Do not crash when loading themed icons statically
We do not officially support static loading of icons.
In fact they still crash for simple png cases due to missing
X11 resources. But since we lazily create themed icons
we can certainly avoid the crash in this case. You will
not be able to use fallbacks here though, since we cannot
know if a fallback should be used or not in this case.

Reviewed-by: ogoffart
Task-number: QTBUG-8666