showFullscreen() broken if UnifiedTitleAndToolBarOnMac is enabled
The problem here is caused by the fix to QTBUG-6200. For this fix we
removed the lines that deleted the item because the iterator might be
invalid at that point. However, we still need to delete the iterator. The
solution is to not use the iterator and ask the hash to delete the item,
since the remove function will not crash if the item is invalid.

Task-number: QTBUG-8633
Reviewed-by: Prasanth