Fix atk warning on startup in Qt apps wit QGtkStyle
While this is acknowledged as a bug in Gtk+ and confirmed
fixed upstream, since some distros such as Ubuntu are not
backporting this fix and it is easy to work around we should
simply fix this for 4.6.3 instead. The problem is simply that
atk assumes the button has a label or a labeled icon and
spits out a warning if it does not.

Additional info can be found here:
https://bugs.launchpad.net/ubuntu/+source/gtk+2.0/+bug/484182

Reviewed-by:thorbjorn
Task-number: QTBUG-8425