Doc: Update QString docs about real ASCII and to/fromAscii and NULs.
Despite the names, the QString::toAscii and QString::fromAscii
functions are not about ASCII. One can set an ASCII-incompatible codec
using QTextCodec::setCodecForCStrings, which may lead to loss of
information.

Also update the docs about how the functions that take QByteArray take
embedded NULs in the QByteArray into account. The result is:

   stop conversion at NUL:      operator=, operator==, operator!=
   include NUL in conversion:   operator<, operator<=, operator>, operator>=

Reviewed-by: ossi