Fixed crash in Symbian when having empty QMenu in menubar.
Typically cascade menus in Symbian/S60 are created from resources.
When cascade menus are defined in resources the RestoreMenuL is
responsible for instantiating the cascade menu item array and items.

In Qt for Symbian the menus need to be created dynamically from
QActions, that's why RestoreMenuL is reimplemeted in Qt for Symbian
to dynamically init the menu pane content. Adding a first dynamic
item to cascase menu pane creates the CEikMenuPane internal item array.
In case where empty QMenu was added to QMenubar, the item array for
cascade menupane was never constructed. Then later on CEikMenuPane code
assumes that item array exist. To avoid access violation crash we now
create the item array manually when building submenu.

Task-number: QTBUG-4897
Reviewed-by: Jason Barron