Set database write behavior to synchronous=OFF and increase page cache.
This improves the INSERT performance dramatically, since we won't wait
anymore until all data is written back to the database. Should not have
any impact on the database consistency in case of an app crash, only in
case the whole OS crashes, but then you might be in trouble anyway.

Reviewed-by: ck