Change WINSCW include paths to use same rules as ARMCC
The default rules applied by symbian cause problems with headers included
from within other headers.
This only affects WINSCW builds, so the headers always need to be specially
fixed up to be compatible, which is wrong.
Instead, use CXXFLAGS to change the include path rules to be the same as

Reviewed-by: Iain