Mac: qfontdialog autotest failes
The reason is the way we leave modality in the font dialog. When we
do, we need to close the non-native qfontdialog as well, but doing
so will issue callbacks (like set_visible_sys(false)), that will
cause problems. This patch makes sure we wait with such callbacks
until the native modal event loop has finished executing

Reviewed-by: cduclos