QPrintPreviewDialog number of pages is partially blocked from view in OSX
The problem here is caused by the FormLayout. This patch queries the
widget for its sizeHint and then forces it to resize to that size.
Once that is done, we add the widget to the layout.
Code is #ifdef protected so other platforms are not affected by it.

Task-number: QTBUG-7760
Reviewed-by: Trond