QMetaType::type(): return immediately if the typename is empty
No need to look up / lock data structures if we know that the
comparisons will all fail.
This was encountered because QMetaMethod::typeName() returns an
empty string if the return type is void (even though there is a
QMetaType::Void (value 0) with name "void"). This was causing
the QtScript meta-object binding to spend a lot of its time
looking up the type for an empty string when invoking slots that
return void.
Rather than having these checks in QtScript and who knows where
else, it's better that QMetaType::type() does it itself.

No regressions in the qmetatype benchmark.

Reviewed-by: Olivier Goffart
Reviewed-by: Harald Fernengel