Prevent a freeze of QFileSystemWatcher on Mac.
On Mac when the FSEvents backend is used and a file is added or removed from a
file system watcher, we need to wait until the thread is finished, otherwise it
is possible that the thread already exited from the run() function but hasn't
fully terminated, meaning when we restart the thread by calling start() it
won't start because QThread thinks it's already running.

A better fix might be to avoid stopping and starting threads - to just stop the
FSEvents loop and notify the thread that a file set has changed.

Task-number: QTBUG-8524
Reviewed-by: Morten