Changed TEST_COMPILER from CC to CXX in configure script.
We are using a cpp file for testing the compiler. Because of that we
have to use a c++ compiler for the test as well.
But if there was a CC variable defined, this one was used instead
of CXX for the compiler test.
This resulted in a missing QMAKE_LFLAGS definition in .qmake.cache.
Due to this, the -L/path/to/QTDIR/lib was missing too and
assistant could not be linked to libQtCLucene.so.4.

Reviewed-by: Marius Storm-Olsen
Task-number: QTBUG-8808,QTBUG-5471,QTBUG-5385,QTBUG-5774