Do not deliver the same key event multiple times in Cocoa.
When QCocoaView receives a keyDown or keyUp event, we translate it to QKeyEvent
and deliver it to qt widgets, propagating if necessary. However if none of Qt
widgets accepted the event we should propagate it back to cocoa - to get a beep
from it, or for example if Qt widget is embedded into a native widget.

Task-number: related to QTBUG-6444
Task-number: QTCREATORBUG-698
Reviewed-by: Richard Moe Gustavsen