Improve scrolling in DirectFB
Since DirectFB always uses the front buffer of a double buffered surface
as the source when blitting we needed to make sure that the front buffer
was valid and hence we called Flip for each scroll. This had negative
performance impacts since you could end up with various parts of a
window being updated at different times.

With this patch we store whether the front-buffer is up to date inside
the window surface and only call Flip in scroll if it's not.

Reviewed-by: Jervey Kong <jervey.kong@nokia.com>